+++ gammapy/analysis/core.py
@@ -6,14 +6,13 @@ from astropy.table import Table
 from regions import CircleSkyRegion
 from gammapy.analysis.config import AnalysisConfig
 from gammapy.cube import FoVBackgroundMaker, MapDatasetMaker, SafeMaskMaker
-from gammapy.datasets import MapDataset, SpectrumDataset
+from gammapy.datasets import MapDataset, SpectrumDataset, FluxPointsDataset
 from gammapy.data import DataStore
 from gammapy.maps import Map, MapAxis, WcsGeom
 from gammapy.modeling import Fit
 from gammapy.datasets import Datasets
 from gammapy.modeling.models import Models
 from gammapy.spectrum import (
-    FluxPointsDataset,
     FluxPointsEstimator,
     ReflectedRegionsBackgroundMaker,
     SpectrumDatasetMaker,
