+++ gammapy/datasets/__init__.py
@@ -1,3 +1,4 @@
 from .core import *
 from .map import *
 from .spectrum import *
+from .flux_points import *
